gamemate-backend/src/requirements.txt
asgiref==3.11.1
Django==6.0.2
django-cors-headers==4.9.0
django-environ==0.13.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
packaging==26.0
psycopg==3.3.3
psycopg-binary==3.3.3
PyJWT==2.11.0
setuptools==82.0.0
sqlparse==0.5.5
wheel==0.46.3
